refactor timeseriesScale into its own file
