+++ frontend/src/metabase/visualizations/lib/timeseries.js
@@ -1,6 +1,5 @@
 /* @flow weak */
 
-import d3 from "d3";
 import moment from "moment-timezone";
 import _ from "underscore";
 
@@ -198,84 +197,6 @@ export function computeTimeseriesTicksInterval(xDomain, xInterval, chartWidth) {
     timeRangeMilliseconds(xDomain),
     maxTicksForChartWidth(chartWidth),
   );
-}
-
-// moment-timezone based d3 scale
-export const timeseriesScale = (
-  { count, interval, timezone },
-  linear = d3.scale.linear(),
-) => {
-  const ms = d =>
-    moment.isMoment(d) ? d.valueOf() : moment.isDate(d) ? d.getTime() : d;
-
-  const s = x => linear(ms(x));
-  s.domain = x => {
-    if (x === undefined) {
-      return linear._domain;
-    }
-    linear._domain = x.map(t => moment(t).tz(timezone));
-    if (interval === "month") {
-      x = wrapValues(ticksForRange(linear._domain, 1), x);
-    }
-    linear.domain(x.map(ms));
-    return s;
-  };
-  s.range = x => {
-    if (x === undefined) {
-      return linear._range;
-    }
-    linear._range = x;
-    if (interval === "month") {
-      const plainScale = d3.scale
-        .linear()
-        .domain(s.domain().map(ms))
-        .range(x);
-      const ticks = ticksForRange(linear._domain, 1);
-      const start = plainScale(ticks[0]);
-      const stop = plainScale(ticks[ticks.length - 1]);
-      const step = (stop - start) / (ticks.length - 1);
-      x = wrapValues(d3.range(ticks.length).map(i => start + i * step), x);
-    }
-    linear.range(x);
-    return s;
-  };
-  s.ticks = () => ticksForRange(s.domain(), count);
-
-  const ticksForRange = ([start, end], count) => {
-    const ticks = [];
-    let tick = start
-      .clone()
-      .tz(timezone)
-      .startOf(interval);
-
-    // We want to use "round" ticks for a given interval (unit). If we're
-    // creating ticks every 50 years, but and the start of the domain is in 1981
-    // we move it be on an even 50-year block. 1981 - (1981 % 50) => 1950;
-    const intervalMod = tick.get(interval);
-    tick.set(interval, intervalMod - (intervalMod % count));
-
-    while (!tick.isAfter(end)) {
-      if (!tick.isBefore(start)) {
-        ticks.push(tick);
-      }
-      tick = tick.clone().add(count, interval);
-    }
-    return ticks;
-  };
-
-  s.copy = () => timeseriesScale({ count, interval, timezone }, linear);
-  d3.rebind(s, linear, "rangeRound", "interpolate", "clamp", "invert");
-  return s;
-};
-
-function wrapValues(x, [start, end]) {
-  if (start < x[0]) {
-    x = [start, ...x];
-  }
-  if (x[x.length - 1] < end) {
-    x = [...x, end];
-  }
-  return x;
 }
 
 // We should always have results_timezone, but just in case we fallback to UTC
